LayoutTests/platform/ios-simulator/fast/dom/custom/document-register-namespace-expected.txt

CONSOLE MESSAGE: line 26: TypeError: undefined is not a function (evaluating 'document.register('html-foo', { prototype: Object.create(HTMLElement.prototype) })')
Using document.register() for extending HTML and non-HTML elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
